session: migrate test-infra to testify for testSessionSerialSuite (#34593)
ref pingcap/tidb#28328
